test: share the AI field marker visual suite between themes
